Better organized the react-query custom fetchers to the urls of the routes provided by Directus
