+++ react/src/constants/directus.constants.ts
@@ -1,1 +1,1 @@
-export const DIRECTUS_URL = 'http://localhost:8055'; // TODO: use enviroment variables to set Directus Url
+export const DIRECTUS_URL = 'http://localhost'; // TODO: use enviroment variables to set Directus Url
